+++ tests/test_info.py
@@ -106,7 +106,7 @@ class TestInfo_03_Online(_TestOOC):
 
         self.c3.ooc('/online')
         self.c3.assert_ooc('Online: {}/{}'.format(7, self.server.config['playerlimit']), over=True)
-        for c in self.server.client_manager.clients:
+        for c in self.server.get_clients():
             if c == self.c3:
                 continue
             c.assert_no_ooc()