+++ tests/test_zonechangewatchers.py
@@ -387,7 +387,8 @@ class TestZoneChangeWatchers_03_Disconnections(_TestZone):
         self.c0.assert_no_packets()
         self.c1.assert_ooc('(X) {} [{}] is no longer watching your zone.'
                            .format(self.c2_dname, 2))
-        self.c1.assert_ooc('All GMs logged out.', over=True)
+        self.c1.assert_ooc('You have logged out the following clients from their GM rank: '
+                           '{} [{}].'.format(self.c2.name, 2), over=True)
         self.c2.assert_ooc('You are no longer a GM.')
         self.c2.assert_packet('FM', None)
         self.c2.assert_ooc('You are no longer watching zone `{}`.'.format('z0'), over=True)