+++ CHANGELOG.md
@@ -552,6 +552,7 @@
 * Added /dump to generate a server dump on request
 * Added /slit, alias to /bloodtrail
 * Clients sending syntactically correct but otherwise unidentifiable packets now silently log to console and server log rather than propagating an uncaught KeyError
+* Allowed /cleargm to take a client ID to log out a particular client from their GM rank
 * Fixed scream_range in area list yaml files not supporting the keyword '<ALL>' to indicate all areas should be able to receive a scream coming from a particular area
 * Fixed scream_range not checking if the areas a scream can reach to from a particular area exist
 * Fixed /scream, /whisper and /party_whisper raising errors if a message was sent to a deafened player with a bypass message starter. They now sent messages but filtered